+++ node/resolvers/catalog/brand.ts
@@ -4,8 +4,8 @@ import { Slugify } from './slug'
 export const resolvers = {
   Brand: {
     active: prop('isActive'),
-    cacheId: brand => Slugify(brand.name),
-    slug: brand => Slugify(brand.name),
+    cacheId: (brand: any) => Slugify(brand.name),
+    slug: (brand: any) => Slugify(brand.name),
     titleTag: prop('title'),
   }
 }