+++ node/resolvers/logistics/index.ts
@@ -7,13 +7,13 @@ export const fieldResolvers = {
 }
 
 export const queries = {
-  logistics: (_, __, {dataSources: {logistics}}: Context) => logistics.shipping(),
+  logistics: (_: any, __: any, {dataSources: {logistics}}: Context) => logistics.shipping(),
 
   nearPickupPoints: (
-    _,
+    _: any,
     { lat, long, maxDistance }: NearPickupPointsArgs,
     {dataSources: {logistics}}: Context
     ) => logistics.nearPickupPoints(lat, long, maxDistance),
 
-  pickupPoint: (_, { id }: PickupPointArgs, {dataSources: {logistics}}: Context) => logistics.pickupById(id),
+  pickupPoint: (_: any, { id }: PickupPointArgs, {dataSources: {logistics}}: Context) => logistics.pickupById(id),
 }