+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.63.5] - 2019-03-29
+### Added
+- `noImplicitAny` flag to `tsconfig.json`.
+
 ## [2.63.4] - 2019-03-28
 
 ## [2.63.3] - 2019-03-28
@@ -21,8 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [2.63.0] - 2019-03-28
 
 ## [2.62.0] - 2019-03-27
-### Added 
-- Add `specificationGroups` in `product`'s resolvers. 
+### Added
+- Add `specificationGroups` in `product`'s resolvers.
 - New mutation `subscribeNewsletter`.
 
 ## [2.61.1] - 2019-03-27